+++ lib/spring/application_manager.rb
@@ -116,7 +116,7 @@ module Spring
     def start_wait_thread(pid, child)
       Process.detach(pid)
 
-      Thread.new {
+      Spring.failsafe_thread {
         # The recv can raise an ECONNRESET, killing the thread, but that's ok
         # as if it does we're no longer interested in the child
         loop do