+++ lib/spring/server.rb
@@ -106,7 +106,7 @@ module Spring
         end
       end
 
-      @applications.values.map { |a| Thread.new { a.stop } }.map(&:join)
+      @applications.values.map { |a| Spring.failsafe_thread { a.stop } }.map(&:join)
     end
 
     def write_pidfile