+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.7.2
+
+* Use `Spring.failsafe_thread` to prevent threads from aborting process due to `Thread.abort_on_exception` when set to `true`
+
 ## 1.7.1
 
 * Specify absolute path to spring binfile when starting the server